pyproject.toml
[tool.poetry]
name = "django-boilerplate"
version = "0.1.0"
description = "django-boilerplate project"
authors = ["violetstair <violetstair@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8, <3.9"
django = "^3.0.7"
djangorestframework = "^3.11.0"
psycopg2-binary = "^2.8.5"
django-extensions = "^2.2.9"
ipython = "^7.15.0"
pyjwt = "^1.7.1"
djangorestframework_simplejwt = "^4.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
